+++ src/VBox/Main/src-server/CaptureScreenSettingsImpl.cpp
@@ -1,4 +1,4 @@
-/* $Id: CaptureScreenSettingsImpl.cpp 75277 2018-11-06 12:29:18Z knut.osmundsen@oracle.com $ */
+/* $Id: CaptureScreenSettingsImpl.cpp 75279 2018-11-06 12:37:11Z andreas.loeffler@oracle.com $ */
 /** @file
  *
  * VirtualBox COM class implementation - Capture settings of one virtual screen.
@@ -42,7 +42,7 @@ struct CaptureScreenSettings::Data
 
     Machine * const                  pMachine;
     ComObjPtr<CaptureScreenSettings> pPeer;
-    unsigned long                    uScreenId;
+    uint32_t                         uScreenId;
 
     // use the XML settings structure in the members for simplicity
     Backupable<settings::CaptureScreenSettings> bd;
@@ -72,7 +72,7 @@ void CaptureScreenSettings::FinalRelease()
  *
  * @returns COM result indicator
  */
-HRESULT CaptureScreenSettings::init(Machine *aParent, unsigned long uScreenId, const settings::CaptureScreenSettings &data)
+HRESULT CaptureScreenSettings::init(Machine *aParent, uint32_t uScreenId, const settings::CaptureScreenSettings& data)
 {
     LogFlowThisFuncEnter();
     LogFlowThisFunc(("aParent: %p\n", aParent));
